Delete extra playgroup file and move game.swift
